the u64 dream ended pretty quickly with 64^3 chunks

+++ src/game/block/chunk_v2.zig
@@ -1,6 +1,6 @@
 wp: worldPosition,
 entity: blecs.ecs.entity_t = 0,
-data: []u64 = undefined,
+data: []u32 = undefined,
 allocator: std.mem.Allocator,
 
 const Chunk = @This();
